+++ TicketingSystem_Client/globalapi.h
