+++ TicketingSystem_Client/images/scissors.png
