+++ TicketingSystem_Client/mainwindow.ui
@@ -7,11 +7,11 @@
     <x>0</x>
     <y>0</y>
     <width>802</width>
-    <height>640</height>
+    <height>663</height>
    </rect>
   </property>
   <property name="windowTitle">
-   <string>TicketingSystem</string>
+   <string>TicketingSystem_Client</string>
   </property>
   <widget class="QWidget" name="centralwidget">
    <widget class="QTextBrowser" name="textBrowser">
