+++ TicketingSystem_Server/main.cpp
@@ -1,4 +1,5 @@
 #include "mainwindow.h"
+
 #include <QApplication>
 
 int main(int argc, char *argv[])
@@ -6,6 +7,5 @@ int main(int argc, char *argv[])
     QApplication a(argc, argv);
     MainWindow w;
     w.show();
-
     return a.exec();
 }
