--- server-client/client/air-tickets.pro
@@ -1,21 +1,0 @@
-#-------------------------------------------------
-#
-# Project created by QtCreator 2019-07-07T17:52:58
-#
-#-------------------------------------------------
-
-QT       += core gui
-
-greaterThan(QT_MAJOR_VERSION, 4): QT += widgets
-
-TARGET = air-tickets
-TEMPLATE = app
-
-
-SOURCES += main.cpp\
-        mainwindow.cpp
-
-HEADERS  += mainwindow.h \
-    globalapi.h
-
-FORMS    += mainwindow.ui
